Added perspective projection and keyboard inputs.
Improved calculations by switching from procedural calculations to matrix multiplications. Now, every shape should inherit from Object, which has a transform matrix and some methods for translations, rotations and scaling.

Also removed any attempts of shading triangle meshes from the code, as it was never the main purpose (although it could be done in the future, but I really doubt I'll have the stomach for it).