Tried to improve inter-system compatibility.
Opted to use windows.h functions by default instead of the ones from conio.h for keyboard inputs, which should (?) make it easier to ensure that other windows systems have the necessary DLLs to run this.

Also tried to read wavefront obj files into graphs, but that went as well as you'd think. Maybe I'll fix it idk.

Finally, made the 'game' aspect of the thing a little better, and we can now jump.