Specifiable resolution and static linkage.
The program allows for the user to specify a resolution at startup so it fits their console window as nicely as they want it to.

Also, added some compilation and linkage flags so that the program gets linked statically, avoiding the need for the distribution of DLLs.

Also moved a translation from outside the setup function into it.