Objects can now be in a parent-children hierarchy.
Each object is now required to have a name (a simple string, doesn't even need to be unique), which should be useful for debugging.

The hierarchy is not dealt with in the most efficient way possible, but whatever, not like I'm running for engine developer of the century here now am I?

Also, Transforms now have a pointer to the object to which they belong.