Scenes and error handling implemented.
Scenes are collections of objects. A scene can have as many objects as desired, as long as two of them do not share a name and/or are the same object. Scenes start out empty, and can be populated by adding objects one by one. When an object is added to a Scene, all of its children are automatically added alongside it, and any attempts to add the children manually afterward will result in a runtime error. Objects can also be removed from a scene by referencing them through their names. Scenes may contain as many cameras as desired, but only one may be selected as the main camera at a time. A way to dynamically change scenes in the engine should be implemented at some point.

The animation engine (which will probably be renamed to just "engine" very soon) is now initialized with a Scene instead of a camera, and has a draw function that draws the entire Scene. If there is no main camera selected within the scene when the draw function is called, a runtime error will be thrown.

Error handling, here, works through standard C++ exceptions. There are currently a few predefined error types, but support will eventually be added for a simple way to throw a quick error message for debugging purposes.

The Camera class has seen some improvements as well: WTC implementation was optimized by not recalculating the matrix every time it is needed, opting instead to store the current WTC in the Camera itself, only updating its value when the Camera's transform changes.

Opted to move the inter-frame delay from the user implementation (which was in the loop function) to the engine implementation, making it a mandatory part of the frame-by-frame dynamics, seen as the presentation is less than ideal without it.

Also
- abstracted away keyboard input to look more like its Unity counterpart,
- added deltaTime because why not,
- moved implementations from their headers into the respective CPP files, and
- general minor improvements to the code.