Added D3D9 wireframe rendering.
It's basic, messy and inefficient, but it's a start, and completely replaces the old console ASCII screen.

Oh, and the X button to close the app doesn't work, since I had to sacrifice the message-receiving capabilities of the window for the sake of the game loop, which means that WM_CLOSE is never handled by the window procedure. On the subject of message passing, waiting on window messages is NOT asynchronous, meaning that user inputs are queued until a message is received, which as you can imagine makes for horrible gameplay, so until I find a better way, this is how it'll be done.

Please try not to use this until it's safer.