Added triangle meshes and the ability to draw them.
Currently, a simple lighting model is implemented, but something weird is happening with the normals; will have to do something about it later.

For now, triangles will be drawn with no shading, using red, green and blue for each vertex and interpolation for the pixels in between.

It's important to note that the drawing processes are very inefficient due to there being one draw call and one host-device data transfer per triangle (and per line, if drawing wireframes). I should make it better as I get more familiarized with Direct3D 9.